+++ src/identify.jl
@@ -39,21 +39,21 @@ function pro(pedigree::Pedigree)
 end
 
 """
-    sibship(pedigree::Pedigree, IDs::Vector{Int64}; halfSibling = true)
+    sibship(pedigree::Pedigree, IDs::Vector{Int}; halfSibling = true)
 
 Return the siblings of specified individuals.
 
 If `halfSibling` is `true`, half-siblings will be included.
 """
-function sibship(pedigree::Pedigree, IDs::Vector{Int64}; halfSibling = true)
-    siblings = Int64[]
+function sibship(pedigree::Pedigree, IDs::Vector{Int}; halfSibling = true)
+    siblings = Int[]
     for ID ∈ IDs
         individual = pedigree[ID]
-        fathers_children = Int64[]
+        fathers_children = Int[]
         if !isnothing(individual.father)
             fathers_children = [child.ID for child ∈ individual.father.children]
         end
-        mothers_children = Int64[]
+        mothers_children = Int[]
         if !isnothing(individual.mother)
             mothers_children = [child.ID for child ∈ individual.mother.children]
         end
@@ -70,21 +70,21 @@ function sibship(pedigree::Pedigree, IDs::Vector{Int64}; halfSibling = true)
 end
 
 """
-    children(pedigree::Pedigree, ID::Int64)
+    children(pedigree::Pedigree, ID::Int)
 
 Return the children of an individual.
 """
-function children(pedigree::Pedigree, ID::Int64)
+function children(pedigree::Pedigree, ID::Int)
     individual = pedigree[ID]
     sort([child.ID for child ∈ individual.children])
 end
 
 """
-    findFounders(pedigree::Pedigree, IDs::Vector{Int64})
+    findFounders(pedigree::Pedigree, IDs::Vector{Int})
 
 Return a vector of founders from whom the `IDs` descend.
 """
-function findFounders(pedigree::Pedigree, IDs::Vector{Int64})
+function findFounders(pedigree::Pedigree, IDs::Vector{Int})
     ancestorIDs = [ancestor(pedigree, ID) for ID ∈ IDs]
     common_ancestorIDs = ∩(ancestorIDs...)
     founderIDs = [ancestorID for ancestorID ∈ common_ancestorIDs
@@ -94,11 +94,11 @@ function findFounders(pedigree::Pedigree, IDs::Vector{Int64})
 end
 
 """
-    _findMRCA(pedigree::Pedigree, probandIDs::Vector{Int64})
+    _findMRCA(pedigree::Pedigree, probandIDs::Vector{Int})
 
 Return the most recent common ancestors of a list of probands.
 """
-function _findMRCA(pedigree::Pedigree, individuals::Vector{Int64})
+function _findMRCA(pedigree::Pedigree, individuals::Vector{Int})
     ancestorIDs = [ancestor(pedigree, ID) for ID ∈ individuals]
     common_ancestorIDs = ∩(ancestorIDs...)
     older_common_ancestorIDs = ancestor(pedigree, common_ancestorIDs)
@@ -107,11 +107,11 @@ function _findMRCA(pedigree::Pedigree, individuals::Vector{Int64})
 end
 
 """
-    _findMinDistanceMRCA(pedigree::Pedigree, individuals::Vector{Int64})
+    _findMinDistanceMRCA(pedigree::Pedigree, individuals::Vector{Int})
 
 Return the minimum distance (meioses) between two individuals.
 """
-function _findMinDistanceMRCA(pedigree::Pedigree, individuals::Vector{Int64})
+function _findMinDistanceMRCA(pedigree::Pedigree, individuals::Vector{Int})
     mrcas = _findMRCA(pedigree, individuals)
     distances = [findDistance(pedigree, individuals, mrca) for mrca ∈ mrcas]
     minimum(distances)
@@ -119,21 +119,21 @@ end
 
 """
     struct GenMatrix
-        individuals::Vector{Int64}
-        ancestors::Vector{Int64}
-        meioses::Matrix{Int64}
+        individuals::Vector{Int}
+        ancestors::Vector{Int}
+        meioses::Matrix{Int}
     end
 
 A matrix that goes with individuals as rows and ancestors as columns.
 """
 struct GenMatrix
-    individuals::Vector{Int64}
-    ancestors::Vector{Int64}
-    meioses::Matrix{Int64}
+    individuals::Vector{Int}
+    ancestors::Vector{Int}
+    meioses::Matrix{Int}
 end
 
 """
-    findMRCA(pedigree::Pedigree, IDs::Vector{Int64})
+    findMRCA(pedigree::Pedigree, IDs::Vector{Int})
 
 Return a [`GenLib.GenMatrix`](@ref) of meioses between individuals and their most recent
 common ancestors (MRCAs).
@@ -150,9 +150,9 @@ pro2 = pro[2]
 genMatrix = gen.findMRCA(ped, [pro1, pro2])
 ```
 """
-function findMRCA(pedigree::Pedigree, individuals::Vector{Int64})
+function findMRCA(pedigree::Pedigree, individuals::Vector{Int})
     mrcaIDs = _findMRCA(pedigree, individuals)
-    meioses_matrix = Matrix{Int64}(undef, length(individuals), length(mrcaIDs))
+    meioses_matrix = Matrix{Int}(undef, length(individuals), length(mrcaIDs))
     for (i, ID) ∈ enumerate(individuals), (j, mrcaID) ∈ enumerate(mrcaIDs)
         meioses_matrix[i, j] = _findMinDistance(pedigree, ID, mrcaID)
     end
@@ -161,13 +161,13 @@ function findMRCA(pedigree::Pedigree, individuals::Vector{Int64})
 end
 
 """
-    ancestor(pedigree::Pedigree, ID::Int64)
+    ancestor(pedigree::Pedigree, ID::Int)
 
 Return a vector of an individual's ancestors.
 """
-function ancestor(pedigree::Pedigree, ID::Int64)
-    ancestorIDs = Set{Int64}()
-    stack = Int64[]
+function ancestor(pedigree::Pedigree, ID::Int)
+    ancestorIDs = Set{Int}()
+    stack = Int[]
     push!(stack, ID)
     while !isempty(stack)
         ID = pop!(stack)
@@ -186,23 +186,23 @@ function ancestor(pedigree::Pedigree, ID::Int64)
 end
 
 """
-    ancestor(pedigree::Pedigree, IDs::Vector{Int64})
+    ancestor(pedigree::Pedigree, IDs::Vector{Int})
 
 Return a vector of several individual's ancestors.
 """
-function ancestor(pedigree::Pedigree, IDs::Vector{Int64})
+function ancestor(pedigree::Pedigree, IDs::Vector{Int})
     ancestorIDs = union([ancestor(pedigree, ID) for ID ∈ IDs]...)
     sort!(ancestorIDs)
 end
 
 """
-    descendant(pedigree::Pedigree, ID::Int64)
+    descendant(pedigree::Pedigree, ID::Int)
 
 Return the descendants of an individual.
 """
-function descendant(pedigree::Pedigree, ID::Int64)
-    descendantIDs = Set{Int64}()
-    stack = Int64[]
+function descendant(pedigree::Pedigree, ID::Int)
+    descendantIDs = Set{Int}()
+    stack = Int[]
     push!(stack, ID)
     while !isempty(stack)
         ID = pop!(stack)
